api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.21.1
httpx==0.28.1
pydantic==2.10.4
python-dotenv==1.0.1
